URL encoding fixes, removed AUR3 code
I've wanted to install the "lv2-c++-tools" package for a long time and
never got around to fixing this script so it would install it for me. I
wrote a little function to URL encode the package names when using the
RPC, so now it should work better.

I also removed all the Aur3.org code. A while ago I added a warning
saying that the site didn't appear to be up to date. It hasn't been
updated since the last version I released, so I just decided to remove
that code.