Removed notification when updated program is running
Most of the time the updated programs that are currently running are
installed via Pacman, so there's little use in telling you that the one
program you installed is currently running when 15 other ones aren't
mentioned. If Pacman did this, then I'd keep this. I liked the idea once
upon a time, but now it seems more like useless code.