Removed --devel option and code
It didn't work all that great and was quite slow since for some you
had to download the whole repository just to get the version number.
And, it is common practice for development packages to be appended
with something like -git but that isn't necessarily required. It was
interesting writing but not worth keeping it here in my opinion.